requirements.txt
apache-airflow
pandas
sqlalchemy
psycopg2-binary
snowflake-connector-python
python-dotenv